requirements.txt
# FaceOff Requirements
# Core dependencies for face swapping application

# Deep Learning & CV
torch>=2.0.0
torchvision>=0.15.0
onnxruntime-gpu>=1.15.0
insightface>=0.7.3
opencv-python>=4.8.0
numpy>=1.24.0

# Image/Video Processing
Pillow>=10.0.0
moviepy>=1.0.3
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# UI
gradio>=3.50.0

# Enhancement Models
realesrgan>=0.3.0
gfpgan>=1.3.8
basicsr>=1.4.2
facexlib>=0.3.0

# Utilities
PyYAML>=6.0
tqdm>=4.66.0
